+++ packages/workflow-engine/src/step/index.ts
@@ -6,3 +6,4 @@
 
 export * from "./create"
 export * from "./execute"
+export * from "./response"